run_tests.sh: be smarter about using the overcommit sysctl
While the counters test has smarts to detect the presence of the
overcommit sysctl, run_tests.sh does not and spits out bash errors on
older kernels. Just skip the setting/restoring of the sysctl if the
kernel doesn't support it.

+++ tests/run_tests.sh
@@ -125,12 +125,16 @@ restore_shm_sysctl() {
 }
 
 setup_dynamic_pool_sysctl() {
-    DYNAMIC_POOL=`cat /proc/sys/vm/nr_overcommit_hugepages`
-    echo 10 > /proc/sys/vm/hugetlb_dynamic_pool
+    if [ -f /proc/sys/vm/nr_overcommit_hugepages ]; then
+    	DYNAMIC_POOL=`cat /proc/sys/vm/nr_overcommit_hugepages`
+    	echo 10 > /proc/sys/vm/hugetlb_dynamic_pool
+    fi
 }
 
 restore_dynamic_pool_sysctl() {
-    echo "$DYNAMIC_POOL" > /proc/sys/vm/nr_overcommit_hugepages
+    if [ -f /proc/sys/vm/nr_overcommit_hugepages ]; then
+        echo "$DYNAMIC_POOL" > /proc/sys/vm/nr_overcommit_hugepages
+    fi
 }
 
 functional_tests () {
